+++ cmd/sessionhub/main.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version   = "0.3.2"
+	version   = "0.3.3"
 	commit    = "none"
 	buildDate = "unknown"
 )